SupprimerHebergementFenetre debug + ajout fenetre de modification

+++ src/Modele/Appartement.java
@@ -8,7 +8,7 @@ public class Appartement extends Hebergement {
     private boolean petitDejeuner;
     private int etage;
 
-    public Appartement(long idHebergement, String nom, String adresse, BigDecimal prixParNuit,
+    public Appartement(int idHebergement, String nom, String adresse, BigDecimal prixParNuit,
                        String description, String specification, List<String> imageUrls,
                        int nombrePieces, boolean petitDejeuner, int etage) {
         super(idHebergement, nom, adresse, prixParNuit, description, specification, imageUrls);
